refact: 파일 이름 및 구조 변경
