+++ DonDog-iOS/DonDog-iOS/Presentation/Features/PostDetail/ViewModels/CommentViewModel.swift
@@ -15,6 +15,10 @@ final class CommentViewModel: ObservableObject {
     private let db = Firestore.firestore()
     private var roomId: String?
     
+    @Published var name: String = "익명"
+    @Published var createdAt: Date = Date()
+    @Published var text: String = ""
+    
     func fetchComments(postId: String) async {
         do {
             let roomId = try await fetchCurrentUserRoomId()
@@ -102,5 +106,38 @@ final class CommentViewModel: ObservableObject {
             .document(comment.commentId)
         
         try await commentRef.delete()
+    }
+    
+    func setCommentData(comment: CommentData) {
+        fetchUserName(of: comment.authorId)
+        createdAt = comment.createdAt.dateValue()
+        text = comment.text
+    }
+    
+    // TODO: User 싱글톤으로 대체
+    private func fetchUserName(of authorId: String) {
+        Firestore.firestore()
+            .collection("Users")
+            .document(authorId)
+            .getDocument { [weak self] snapshot, error in
+                guard let self = self else { return }
+                
+                if let error = error {
+                    print("사용자 이름을 가져오지 못했습니다: \(error.localizedDescription)")
+                    return
+                }
+                
+                if let data = snapshot?.data(),
+                   let name = data["name"] as? String {
+                    DispatchQueue.main.async {
+                        self.name = name
+                    }
+                } else {
+                    print("사용자 이름이 존재하지 않습니다.")
+                    DispatchQueue.main.async {
+                        self.name = "익명"
+                    }
+                }
+            }
     }
 }
