+++ DonDog-iOS/DonDog-iOS/Presentation/Features/PostDetail/Views/SubViews/CommentView.swift
@@ -15,22 +15,39 @@ struct CommentView: View {
     var body: some View {
         VStack(spacing: 0) {
             ForEach(viewModel.comments, id: \.createdAt) { comment in
-                CommentRow(comment: comment)
-                    .contextMenu {
-                        Button(role: .destructive) {
-                            Task {
-                                do {
-                                    try await viewModel.deleteComment(for: comment, in: postId)
-                                    shouldBeUpdated = true
-                                } catch {
-                                    print("댓글 삭제에 실패했습니다: \(error)")
-                                }
-                            }
-                        } label: {
-                            Text("삭제")
-                            Image(systemName: "trash")
-                        }
+                VStack(alignment: .leading, spacing: 4) {
+                    HStack(spacing: 4) {
+                        Text(viewModel.name)
+                            .font(.captionMedium14)
+                        Text(DateUtils.relativeTimeString(from: viewModel.createdAt))
+                            .font(.captionRegular13)
+                            .foregroundStyle(Color.ddGray600)
                     }
+                    Text(viewModel.text)
+                        .font(.captionRegular14)
+                }
+                .frame(maxWidth: .infinity, alignment: .leading)
+                .padding(.horizontal, 20)
+                .padding(.vertical, 8)
+                .contentShape(Rectangle())
+                .onAppear {
+                    viewModel.setCommentData(comment: comment)
+                }
+                .contextMenu {
+                    Button(role: .destructive) {
+                        Task {
+                            do {
+                                try await viewModel.deleteComment(for: comment, in: postId)
+                                shouldBeUpdated = true
+                            } catch {
+                                print("댓글 삭제에 실패했습니다: \(error)")
+                            }
+                        }
+                    } label: {
+                        Text("삭제")
+                        Image(systemName: "trash")
+                    }
+                }
             }
         }
         .task(id: shouldBeUpdated) {
