--- DonDog-iOS/DonDog-iOS/Presentation/Features/PostDetail/Views/SubViews/PostDetailContentView.swift
@@ -1,19 +1,0 @@
-//
-//  PostDetailContentView.swift
-//  DonDog-iOS
-//
-//  Created by 이서현 on 10/26/25.
-//
-
-import SwiftUI
-
-struct PostDetailContentView: View {
-    let postType: PostType
-    let post: PostData
-    @Binding var shouldBeUpdated: Bool
-    
-    var body: some View {
-        CardView(post: post)
-        CommentView(postId: post.postId, shouldBeUpdated: $shouldBeUpdated)
-    }
-}
